Fix bug966. Assertion in cgexpand under m32 O2.
case:
typedef unsigned char uint8_t;
typedef unsigned int uint32_t;
typedef unsigned long long int uint64_t;
extern memddd(unsigned int);
void build_avp(uint32_t avp_vendor, uint64_t in_value_len)
{
  if (avp_vendor != 0)
    {
      in_value_len = in_value_len +4;
    }
    {
      memddd(in_value_len);
    }
}
The assertion failure is triggered during Code_Expansion phase, when expanding I8SELECT, the pair TN of its kid is null, so it hit the assertion. But the whirl IR after WOPT phase was not correct, so the root cause should be in WOPT phase. In SELECT, Both Kid 1 and Kid 2 must have res as the result type. For ex, in above case, the Dtyp of kid 1 turn to U4 from U8 in BITWISE_DCE phase. So the fix method here is to generate CVT for the kids to make sure their Dtyp are the same with SELECT.

The patch handle following situations:
1)   LDC I8      2)   LDC U4         3)   LDC U4          4)     LDC U4      5)   LDC U4
     LDC I8           LDC U4              LDC U4                 LDC I8           LDC U4
   I8SELECT         U4SELECT            I8SELECT               I8SELECT         I4SELECT
for 1) and 2), it is of right form;
for 3) and 5), we will set the type of SELECT to U4;
for 4), I8U4CVT will be added for U4.

Author of this patch:Xiaojing-zhang

Code Review: Sun chan


git-svn-id: https://svn.open64.net/svnroot/open64/trunk@3973 ead328b1-e903-0410-a19d-eacee4b031f6